FLASK_TUTORIAL/requirements.txt
flask
pymango
dnspython
python-dotenv